Remove RestorationImageCache
