+++ Modules/Legacy/Core/Sources/Application/StateRestorationError.swift
@@ -1,0 +1,6 @@
+//  Created by Geoff Pado on 6/24/24.
+//  Copyright © 2024 Cocoatype, LLC. All rights reserved.
+
+enum StateRestorationError: Error {
+    case invalidImageData
+}
